Add live probe + consolidate SystemUser install helpers
Add a `probe(session): boolean` to `ServerSupportFeature`, a live
presence check distinct from the cached `isMissing` flag, so future
test gating can ask "is this actually installed right now" instead of
relying on the availability latch set at connect time.

Also extract the three-way-duplicated install helpers
(`gemCanRead`/`gsStringLiteral`/`safeAbort`/`messageOf`/
`yieldToEventLoop`/`loginAsSystemUser`) into `client/src/serverPlugin/`,
split into a `vscode`-free `installHelpers.ts` and a
`systemUserAuth.ts` for the interactive password prompt. The CI
provisioning script now imports `loginAsSystemUser` from there instead
of keeping its own copy.
