+++ client/src/__tests__/extension.test.ts
@@ -24,6 +24,7 @@ vi.mock('../browserQueries', () => ({
     }
   },
   compileMethod: vi.fn(() => 'Compiled: Array >> foo'),
+  checkRefactoringSupportAvailable: vi.fn(() => false),
 }));
 
 import * as vscode from 'vscode';
