+++ client/src/refactoring/refactoringInstall.ts
@@ -34,6 +34,13 @@
 import { ActiveSession } from '../sessionManager';
 import { executeFetchString, checkRefactoringSupportAvailable } from '../browserQueries';
 import { compareGemStoneVersions } from '../gemStoneVersion';
+import {
+  gemCanRead,
+  gsStringLiteral,
+  messageOf,
+  safeAbort,
+  yieldToEventLoop,
+} from '../serverPlugin/installHelpers';
 
 /**
  * The bootstrap file the client files in itself: it defines `GsRefactoringLoader`
@@ -211,43 +218,4 @@ export function supportsServerUtf8FileIn(stoneVersion: string | undefined): bool
   } catch {
     return false;
   }
-}
-
-/** Whether the gem process can read the file at `serverPath`. */
-function gemCanRead(session: ActiveSession, serverPath: string): boolean {
-  try {
-    const r = executeFetchString(
-      session,
-      `[(GsFile existsOnServer: ${gsStringLiteral(serverPath)}) printString] ` +
-        "on: Error do: [:e | 'false']",
-    );
-    return r.trim() === 'true';
-  } catch {
-    return false;
-  }
-}
-
-/** Render a JS string as a GemStone string literal: single quotes doubled and
- *  the whole value wrapped in quotes. */
-function gsStringLiteral(s: string): string {
-  return `'${s.replace(/'/g, "''")}'`;
-}
-
-/** Yield to the event loop so the progress notification can paint between the
- *  (synchronous) server calls. */
-function yieldToEventLoop(): Promise<void> {
-  return new Promise<void>((resolve) => setImmediate(resolve));
-}
-
-function safeAbort(session: ActiveSession): void {
-  try {
-    session.gci.GciTsAbort(session.handle);
-  } catch {
-    // Best-effort rollback; the caller closes the session regardless.
-  }
-}
-
-/** Extract a human-readable message from a thrown value. */
-export function messageOf(e: unknown): string {
-  return e instanceof Error ? e.message : String(e);
 }
